python/requirements.txt
dnspython==2.4.2
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
pymongo==4.5.0
